backend/requirements.txt
flask
flask-cors
openai
requests
pandas
scikit-learn
joblib
python-dotenv
